+++ Services/Iam/V3/Model/CreateUnscopeTokenByIdpInitiatedRequest.cs
@@ -8,20 +8,27 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 using HuaweiCloud.SDK.Core;
 
-namespace HuaweiCloud.SDK.Moderation.V3.Model
+namespace HuaweiCloud.SDK.Iam.V3.Model
 {
     /// <summary>
     /// Request Object
     /// </summary>
-    public class RunCreateVideoStreamModerationJobRequest 
+    public class CreateUnscopeTokenByIdpInitiatedRequest 
     {
+
+        /// <summary>
+        /// 身份提供商ID。
+        /// </summary>
+        [SDKProperty("X-Idp-Id", IsHeader = true)]
+        [JsonProperty("X-Idp-Id", NullValueHandling = NullValueHandling.Ignore)]
+        public string XIdpId { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
         [SDKProperty("body", IsBody = true)]
         [JsonProperty("body", NullValueHandling = NullValueHandling.Ignore)]
-        public VideoStreamCreateRequest Body { get; set; }
+        public CreateUnscopeTokenByIdpInitiatedRequestBody Body { get; set; }
 
 
 
@@ -31,7 +38,8 @@ namespace HuaweiCloud.SDK.Moderation.V3.Model
         public override string ToString()
         {
             var sb = new StringBuilder();
-            sb.Append("class RunCreateVideoStreamModerationJobRequest {\n");
+            sb.Append("class CreateUnscopeTokenByIdpInitiatedRequest {\n");
+            sb.Append("  xIdpId: ").Append(XIdpId).Append("\n");
             sb.Append("  body: ").Append(Body).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
@@ -42,15 +50,16 @@ namespace HuaweiCloud.SDK.Moderation.V3.Model
         /// </summary>
         public override bool Equals(object input)
         {
-            return this.Equals(input as RunCreateVideoStreamModerationJobRequest);
+            return this.Equals(input as CreateUnscopeTokenByIdpInitiatedRequest);
         }
 
         /// <summary>
         /// Returns true if objects are equal
         /// </summary>
-        public bool Equals(RunCreateVideoStreamModerationJobRequest input)
+        public bool Equals(CreateUnscopeTokenByIdpInitiatedRequest input)
         {
             if (input == null) return false;
+            if (this.XIdpId != input.XIdpId || (this.XIdpId != null && !this.XIdpId.Equals(input.XIdpId))) return false;
             if (this.Body != input.Body || (this.Body != null && !this.Body.Equals(input.Body))) return false;
 
             return true;
@@ -64,6 +73,7 @@ namespace HuaweiCloud.SDK.Moderation.V3.Model
             unchecked // Overflow is fine, just wrap
             {
                 var hashCode = 41;
+                if (this.XIdpId != null) hashCode = hashCode * 59 + this.XIdpId.GetHashCode();
                 if (this.Body != null) hashCode = hashCode * 59 + this.Body.GetHashCode();
                 return hashCode;
             }